+++ src/core/async/thread.cpp
@@ -142,28 +142,23 @@ void* thread::thread_entry(void* arg) {
 void thread::start_thread_impl(thread_startup_args* args) {
     hook::invoke(hook::point::before_create, id_);
 
+    id::native_id_type tid = 0;
+
 #ifdef NEFORCE_PLATFORM_WINDOWS
-    // NOLINTNEXTLINE(cppcoreguidelines-init-variables)
-    unsigned int thread_id;
-    handle_ = reinterpret_cast<native_handle_type>(::_beginthreadex(nullptr, 0, thread_entry, args, 0, &thread_id));
+    handle_ = reinterpret_cast<native_handle_type>(::_beginthreadex(nullptr, 0, thread_entry, args, 0, &tid));
     if (handle_ == nullptr) {
-        // NOLINTNEXTLINE(cppcoreguidelines-owning-memory)
         delete args;
         NEFORCE_THROW_EXCEPTION(thread_exception("Failed to create thread"));
     }
-    id_ = id(thread_id);
-
 #else
-    native_handle_type tid = 0;
     if (::pthread_create(&tid, nullptr, thread_entry, args) != 0) {
         delete args;
         NEFORCE_THROW_EXCEPTION(thread_exception("Failed to create thread"));
     }
-    handle_ = tid;
-    id_ = id(tid);
-
+    handle_ = static_cast<native_handle_type>(tid);
 #endif
 
+    id_ = id(tid);
     state_ = CREATED;
     hook::invoke(hook::point::after_create, id_);
 }